Smooth movement and logic in proper classes
This took me a minute, but I finally got smooth jumping that properly checks when the player is able to jump. I also added where there is less directional control when the player is in the air. He still collides with the sides of platforms but I figured we'd fix that a different issue since it's not actually part of movement. Let me know if you find any bugs.